refactor: distribute layout exports to subdirectory index files
- Remove main src/components/layout/index.ts
- Create wrappers/index.ts for wrapper component exports
- Create common/index.ts for common component exports
- Update all app imports to use specific subdirectory paths
- Fix unused import warning in settings.tsx
