+++ src/app/settings.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { Switch, Pressable } from 'react-native';
-import { router } from 'expo-router';
-import { ModalWrapper } from '@/components/layout';
+import { ModalWrapper } from '@/components/layout/wrappers';
 import { ThemedView } from '@/components/themed/themed-view';
 import { ThemedText } from '@/components/themed/themed-text';
 import { IconSymbol, IconSymbolName } from '@/components/ui/icon-symbol';
