--- src/components/layout/index.ts
@@ -1,12 +1,0 @@
-// Wrappers
-export { ScreenWrapper } from './wrappers/screen-wrapper';
-export type { ScreenWrapperProps } from './wrappers/screen-wrapper';
-export { TabScreenWrapper } from './wrappers/tab-screen-wrapper';
-export type { TabScreenWrapperProps } from './wrappers/tab-screen-wrapper';
-export { ModalWrapper } from './wrappers/modal-wrapper';
-
-// Common Components
-export { SafeAreaWrapper } from './wrappers/safe-area-wrapper';
-export { LoadingState } from './common/loading-state';
-export { ErrorState } from './common/error-state';
-export { EmptyState } from './common/empty-state';
